Safari-611.1.10/LayoutTests/http/tests/security/cross-frame-access-object-prototype-expected.txt


SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.

----- test getting values cross-frame using Object.prototype -----

PASS Object.prototype.__lookupGetter__.call(targetFrame, 'myGetter'); threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS Object.prototype.__lookupSetter__.call(targetFrame, 'mySetter'); threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS Object.prototype.propertyIsEnumerable.call(targetFrame, 'myProp'); threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS Object.prototype.hasOwnProperty.call(targetFrame, 'myProp'); threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..


--------
Frame: 'target'
--------

----- test that functions could not be injected cross-frame using Object.prototype -----

PASS: window.hasOwnProperty('injectedGetter') should be 'false' and is.
PASS: window.__lookupGetter__('injectedGetter') should be 'undefined' and is.
PASS: window.hasOwnProperty('injectedSetter') should be 'false' and is.
PASS: window.__lookupSetter__('injectedSetter') should be 'undefined' and is.
PASS: window.injectedGetter should be 'undefined' and is.
PASS: window.injectedSetter = 1 should be '1' and is.
